+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 
 ##### Enhancements
 
-* None.
+* Support using pre-generated symbolgraph files in Swift symbolgraph mode.
+  [Nathan Wong](https://github.com/esteluk)
 
 ##### Bug Fixes